requirements.txt
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
exceptiongroup==1.2.2
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
idna==3.10
iniconfig==2.0.0
multidict==6.1.0
numpy==2.0.2
packaging==24.2
pluggy==1.5.0
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pytest==7.4.3
pytest-recording==0.13.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
six==1.17.0
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vcrpy==7.0.0
wrapt==1.17.2
yarl==1.18.3
